requirements.txt
# Core dependencies
yt-dlp>=2023.12.30
requests>=2.31.0
python-dotenv>=1.0.0

# Video processing
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
pillow>=10.0.0

# Audio processing
openai-whisper>=20231117

# AI/ML
openai>=1.6.0

# CLI and utilities
click>=8.1.0
tqdm>=4.66.0
colorama>=0.4.6

# Data handling
pydantic>=2.5.0

# Optional: For better image handling
imageio>=2.33.0
imageio-ffmpeg>=0.4.9